+++ CHANGELOG.md
@@ -10,7 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 * Enable `aabb` feature by default
-* Ability to offset an `Aabb` shape
+* `Aabb::offset` and `Aabb::with_offset`
+* `Aabb::size`
 
 
 ## [0.2.1] - 2024-09-01